hexh: Split oversized modules and regenerate workspace-hack
