+++ server/crates/djinn-cgroup-launcher/Cargo.toml
@@ -10,6 +10,7 @@ doctest = false
 [dependencies]
 libc = "0.2"
 thiserror = "2"
+workspace-hack.workspace = true
 
 [lints]
 workspace = true
