+++ server/crates/djinn-slot/src/lib.rs
@@ -43,6 +43,7 @@ pub mod attempt_lifecycle;
 pub mod commands;
 pub mod extraction_replay_eval;
 pub mod final_verification;
+pub mod final_verification_agent;
 pub mod finalize_handlers;
 pub mod finalize_types;
 pub mod helpers;
